requirements.txt
fastapi>=0.110
uvicorn>=0.27
requests>=2.31
feedparser>=6.0
beautifulsoup4>=4.12
lxml>=5.1
python-dotenv>=1.0
google-genai>=0.3
reportlab>=4.0